fintech/requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
langchain-community>=0.0.1
langchain-anthropic>=0.0.1
langchain-openai>=0.0.1
firebase-functions>=0.1.0
firebase-admin>=6.5.0
openai>=1.61.0
pandas>=2.1.4
langchain-community>=0.0.10
langchain-core>=0.1.10
google-cloud-firestore>=2.10.0
openai>=1.12.0
anthropic>=0.18.1
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.36
tqdm>=4.66.1 
tiktoken>=0.4.0